nvim: wip / setup which key for all the functions
This is not done at all quite yet. I want to have all the bindings
follow the same pattern. I am not sure yet if I want to have things like
go to definition under the leader key too or if I should leave it
without. If I do leave it outside, how do I make sense of it so there is
still a pattern I can make sense of without consulting the config files
all the time?